requirements.txt
aenum==2.2.4
blinker==1.3
cassandra-driver==3.24.0
chardet==2.3.0
click==7.1.2
cloud-init==20.3
command-not-found==0.3
configobj==5.0.6
cryptography==1.2.3
decorator==4.0.6
Django==2.2.16
gensim==3.8.3
geomet==0.1.2
gremlinpython==3.4.6
idna==2.0
ipython==2.4.1
isodate==0.6.0
Jinja2==2.8
joblib==0.14.1
jsonpatch==1.10
jsonpointer==1.9
language-selector==0.1
MarkupSafe==0.23
nltk==3.5
numpy==1.18.5
oauthlib==1.0.3
prettytable==0.7.2
pyasn1==0.1.9
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
pyserial==3.0.1
python-apt==1.1.0b1+ubuntu0.16.4.9
python-debian==0.1.27
python-systemd==231
pytz==2020.1
PyYAML==3.11
regex==2020.10.11
requests==2.9.1
scipy==1.4.1
simplegeneric==0.8.1
six==1.10.0
smart-open==3.0.0
sqlparse==0.4.1
ssh-import-id==5.5
tornado==4.5.3
tqdm==4.50.2
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1

